+++ CHANGELOG.md
@@ -9,9 +9,12 @@ Unreleased changes are in the `master` branch.
 
 ## [Unreleased]
 
+## [0.1.57] - 2023-03-17
+
 ### Fixed
 
-- `getSchemaForEndpoint` compatibility with fast-json-stringify (array `type` not supported except for `["<type>", "null"]`)
+- `getSchemaForEndpoint` compatibility with fast-json-stringify (array in `type` not supported except for `["<type>", "null"]`, fix for nested arbitrary objects)
+- generated ts types and schema for `/ipfs/gateway/{IPFS_path}`
 
 ## [0.1.56] - 2023-03-15